[arcane,accelerator] Déplace les implémentations des algorithmes dans le fichier de même nom que la classe correspondante.
